+++ src/lib/singleview/general/docmodel.test.ts
@@ -7,7 +7,7 @@
 // Software-Engineering: 2023 Intevation GmbH <https://intevation.de>
 
 import { describe, it, expect } from "vitest";
-import { convertToDocModel } from "$lib/singleview/docmodel/docmodel";
+import { convertToDocModel } from "$lib/singleview/general/docmodel";
 import {
   EMPTY,
   Status,
@@ -15,7 +15,7 @@ import {
   type DocModel,
   type DocModelKey,
   TLP
-} from "$lib/singleview/docmodel/docmodeltypes";
+} from "$lib/singleview/general/docmodeltypes";
 
 const allEmpty = (docModel: DocModel, properties: DocModelKey[]) => {
   properties.forEach((p) => {
