+++ src/lib/singleview/vulnerabilities/Vulnerabilities.svelte
